// src/types/enums.ts
export enum RiskLevel {
  HIGH = "High",
  MEDIUM = "Medium",
  LOW = "Low",
  CRITICAL = "Critical",
}

export enum PaymentMethod {
  CREDIT_CARD = "Credit Card",
  BANK_TRANSFER = "Bank Transfer",
  ELECTRONIC_CHECK = "Electronic Check",
  MAILED_CHECK = "Mailed Check",
  DEBIT_CARD = "Debit Card",
}

export enum SegmentType {
  CHAMPIONS = "Champions",
  LOYAL_CUSTOMERS = "Loyal Customers",
  POTENTIAL_LOYALISTS = "Potential Loyalists",
  AT_RISK = "At Risk",
  CANNOT_LOSE_THEM = "Cannot Lose Them",
}

export enum RFMSegmentType {
  BEST = "Best",
  LOYAL = "Loyal",
  BIG_SPENDER = "Big Spender",
  NEW = "New",
  PROMISING = "Promising",
  LOST_POTENTIAL = "Lost Potential",
  LOST = "Lost",
}

export enum ProductCategory {
  FASHION = "Fashion",
  LAPTOP_ACCESSORY = "Laptop & Accessory",
  MOBILE_PHONE = "Mobile Phone",
}